refactor: make some minor changes in events
