+++ ewm/src/main/java/ru/practicum/ewm/event/EventMapper.java
@@ -7,6 +7,7 @@ import ru.practicum.ewm.event.dto.EventFullDto;
 import ru.practicum.ewm.event.dto.EventShortDto;
 import ru.practicum.ewm.event.dto.NewEventDto;
 import ru.practicum.ewm.event.model.Event;
+import ru.practicum.ewm.event.state.EventState;
 import ru.practicum.ewm.user.UserMapper;
 import ru.practicum.ewm.user.model.User;
