+++ ewm/src/main/java/ru/practicum/ewm/event/dto/EventFullDto.java
@@ -5,7 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import ru.practicum.ewm.event.EventState;
+import ru.practicum.ewm.event.state.EventState;
 import ru.practicum.ewm.event.model.Location;
 import ru.practicum.ewm.category.dto.CategoryDto;
 import ru.practicum.ewm.user.dto.UserShortDto;
