+++ ewm/src/main/java/ru/practicum/ewm/event/model/Event.java
@@ -5,7 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import ru.practicum.ewm.category.model.Category;
-import ru.practicum.ewm.event.EventState;
+import ru.practicum.ewm.event.state.EventState;
 import ru.practicum.ewm.user.model.User;
 
 import javax.persistence.*;
