+++ ewm/src/main/java/ru/practicum/ewm/event/service/EventPrivateServiceImpl.java
@@ -9,8 +9,8 @@ import ru.practicum.ewm.category.dto.CategoryDto;
 import ru.practicum.ewm.category.model.Category;
 import ru.practicum.ewm.category.service.CategoryPublicService;
 import ru.practicum.ewm.event.EventMapper;
-import ru.practicum.ewm.event.EventState;
-import ru.practicum.ewm.event.ParticipationStatus;
+import ru.practicum.ewm.event.state.EventState;
+import ru.practicum.ewm.event.state.ParticipationStatus;
 import ru.practicum.ewm.event.dto.EventFullDto;
 import ru.practicum.ewm.event.dto.EventShortDto;
 import ru.practicum.ewm.event.dto.NewEventDto;
