+++ ewm/src/main/java/ru/practicum/ewm/event/state/ParticipationStatus.java
@@ -1,4 +1,4 @@
-package ru.practicum.ewm.event;
+package ru.practicum.ewm.event.state;
 
 public enum ParticipationStatus {
