+++ ewm/src/main/java/ru/practicum/ewm/event/state/SortOption.java
@@ -1,0 +1,17 @@
+package ru.practicum.ewm.event.state;
+
+public enum SortOption {
+
+    EVENT_DATE, // Сортировка по дате события
+
+    VIEWS;      // Сортировка по количеству просмотров
+
+    public static SortOption from(String state) {
+        for (SortOption value : SortOption.values()) {
+            if (value.name().equals(state)) {
+                return value;
+            }
+        }
+        return null;
+    }
+}
