+++ ewm/src/main/java/ru/practicum/ewm/request/model/ParticipationRequest.java
@@ -4,7 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import ru.practicum.ewm.event.ParticipationStatus;
+import ru.practicum.ewm.event.state.ParticipationStatus;
 import ru.practicum.ewm.event.model.Event;
 import ru.practicum.ewm.user.model.User;
