+++ ewm/src/main/java/ru/practicum/ewm/request/service/ParticipationRequestPrivateServiceImpl.java
@@ -3,9 +3,9 @@ package ru.practicum.ewm.request.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import ru.practicum.ewm.event.EventState;
+import ru.practicum.ewm.event.state.EventState;
 import ru.practicum.ewm.event.EventValidator;
-import ru.practicum.ewm.event.ParticipationStatus;
+import ru.practicum.ewm.event.state.ParticipationStatus;
 import ru.practicum.ewm.event.model.Event;
 import ru.practicum.ewm.exception.EntityAlreadyExistsException;
 import ru.practicum.ewm.exception.ValidationException;
